Refactor comment and feed entities

+++ lib/feed/logic/bloc/feed_bloc.dart
@@ -138,42 +138,33 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
   }
 
   Future<void> _onCreatePost(CreatePost event, Emitter<FeedState> emit) async {
-    if (state is! FeedLoaded) return;
+    if (state is! FeedLoaded || state is PostCreating) return;
     final currentState = state as FeedLoaded;
-    if (currentState.isCreating) return;
     try {
-      emit(currentState.copyWith(isCreating: true, isCreateSuccess: null));
+      emit(PostCreating());
       final createdPost = await feedRepository.createPost(
         CreatePostRequest(text: event.text, media: event.media),
       );
+      emit(PostCreated());
       emit(
         currentState.copyWith(
           posts: [createdPost.post, ...currentState.posts],
-          isCreating: false,
-          createError: null,
           media: null,
-          isCreateSuccess: true,
         ),
       );
     } catch (e, st) {
-      emit(
-        currentState.copyWith(
-          isCreating: false,
-          createError: e.toString(),
-          isCreateSuccess: false,
-        ),
-      );
+      emit(PostCreatingFailure(error: e));
+      emit(currentState);
       talker.handle(e, st);
       errorHandler.handle(e);
     }
   }
 
   Future<void> _onUpdatePost(EditPost event, Emitter<FeedState> emit) async {
-    if (state is! FeedLoaded) return;
+    if (state is! FeedLoaded || state is PostUpdating) return;
 
     final currentState = state as FeedLoaded;
-    if (currentState.isUpdating) return;
-    emit(currentState.copyWith(isUpdating: true, isUpdateSuccess: false));
+    emit(PostUpdating());
     try {
       final response = await feedRepository.editPost(
         EditPostRequest(
@@ -188,27 +179,27 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
           .where((post) => post.id != event.postId)
           .toList();
 
+      emit(PostUpdated());
+
       emit(
         currentState.copyWith(
           posts: [response.post, ...updatedPosts],
-          isUpdateSuccess: true,
-          isUpdating: false,
           media: null,
         ),
       );
     } catch (e, st) {
-      emit(currentState.copyWith(isUpdateSuccess: false, isUpdating: false));
+      emit(PostUpdatingFailure(error: e));
+      emit(currentState);
       talker.handle(e, st);
       errorHandler.handle(e);
     }
   }
 
   Future<void> _onDeletePost(DeletePost event, Emitter<FeedState> emit) async {
-    if (state is! FeedLoaded) return;
+    if (state is! FeedLoaded || state is PostDeleting) return;
 
     final currentState = state as FeedLoaded;
-    if (currentState.isDeleting) return;
-    emit(currentState.copyWith(isDeleting: true, isDeleteSuccess: false));
+    emit(PostDeleting());
     try {
       final response = await feedRepository.deletePost(
         DeletePostRequest(postId: event.postId),
@@ -224,22 +215,19 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
           .where((post) => post.id != event.postId)
           .toList();
 
-      emit(
-        currentState.copyWith(
-          posts: updatedPosts,
-          isDeleteSuccess: true,
-          isDeleting: false,
-        ),
-      );
+      emit(PostDeleted());
+
+      emit(currentState.copyWith(posts: updatedPosts));
     } catch (e, st) {
-      emit(currentState.copyWith(isDeleteSuccess: false, isDeleting: false));
+      emit(PostDeletingFailure(error: e));
+      emit(currentState);
       talker.handle(e, st);
       errorHandler.handle(e);
     }
   }
 
   Future<void> _onToggleLike(ToggleLike event, Emitter<FeedState> emit) async {
-    if (state is! FeedLoaded) return;
+    if (state is! FeedLoaded || state is PostLiking) return;
     final currentState = state as FeedLoaded;
 
     final targetPost = currentState.posts.firstWhere(
@@ -262,7 +250,7 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
       return post;
     }).toList();
 
-    emit(currentState.copyWith(posts: optimisticPosts, likeError: null));
+    emit(PostLiking());
 
     try {
       final response = willLike
@@ -277,23 +265,17 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
           originalPost: originalPost,
           currentState: currentState,
         );
-        emit(
-          currentState.copyWith(
-            posts: rolledBackPosts,
-            likeError: errorMessage,
-          ),
-        );
+        emit(currentState.copyWith(posts: rolledBackPosts));
         throw ApiException(message: errorMessage);
       }
     } catch (e, st) {
+      emit(PostLikingFailure(error: e));
       final rolledBackPosts = _getRolledBackPosts(
         postId: event.postId,
         originalPost: originalPost,
         currentState: currentState,
       );
-      emit(
-        currentState.copyWith(posts: rolledBackPosts, likeError: errorMessage),
-      );
+      emit(currentState.copyWith(posts: rolledBackPosts));
       talker.handle(e, st);
       errorHandler.handle(e);
     }
@@ -316,15 +298,13 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
     PickImageFromCamera event,
     Emitter<FeedState> emit,
   ) async {
-    if (state is! FeedLoaded) return;
+    if (state is! FeedLoaded || state is MediaLoading) return;
     final currentState = state as FeedLoaded;
-
-    emit(currentState.copyWith(isMediaLoading: true));
-
+    emit(MediaLoading());
     try {
       final image = await mediaService.takePhotoWithPermission();
       if (image == null) {
-        emit(currentState.copyWith(isMediaLoading: false));
+        emit(currentState);
         return;
       }
       final currentMedia = currentState.media ?? [];
@@ -332,18 +312,13 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
 
       if (totalCount > 10) {
         errorHandler.handle("Можно выбрать максимум 10 файлов");
-        emit(currentState.copyWith(isMediaLoading: false));
+        emit(currentState);
         return;
       }
 
-      emit(
-        currentState.copyWith(
-          isMediaLoading: false,
-          media: [image, ...currentMedia],
-        ),
-      );
+      emit(currentState.copyWith(media: [image, ...currentMedia]));
     } catch (e, st) {
-      emit(currentState.copyWith(isMediaLoading: false));
+      emit(currentState);
       talker.handle(e, st);
       errorHandler.handle(e);
     }
@@ -353,16 +328,16 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
     PickMediaFromGallery event,
     Emitter<FeedState> emit,
   ) async {
-    if (state is! FeedLoaded) return;
+    if (state is! FeedLoaded || state is MediaLoading) return;
     final currentState = state as FeedLoaded;
 
-    emit(currentState.copyWith(isMediaLoading: true));
+    emit(MediaLoading());
 
     try {
       final media = await mediaService.pickMultipleMedia();
 
       if (media.isEmpty) {
-        emit(currentState.copyWith(isMediaLoading: false));
+        emit(currentState);
         return;
       }
 
@@ -371,7 +346,7 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
 
       if (totalCount > 10) {
         errorHandler.handle("Можно выбрать максимум 10 файлов");
-        emit(currentState.copyWith(isMediaLoading: false));
+        emit(currentState);
         return;
       }
 
@@ -387,18 +362,13 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
         errorHandler.handle(
           "Не удалось прикрепить файл. Размер каждого файла должен быть не больше 10 МБ",
         );
-        emit(currentState.copyWith(isMediaLoading: false));
+        emit(currentState);
         return;
       }
 
-      emit(
-        currentState.copyWith(
-          isMediaLoading: false,
-          media: [...filteredMedia, ...currentMedia],
-        ),
-      );
+      emit(currentState.copyWith(media: [...filteredMedia, ...currentMedia]));
     } catch (e, st) {
-      emit(currentState.copyWith(isMediaLoading: false));
+      emit(currentState);
       talker.handle(e, st);
       errorHandler.handle(e);
     }
@@ -453,9 +423,12 @@ class FeedBloc extends Bloc<FeedEvent, FeedState> {
           )
           .toList();
 
-      emit(current.copyWith(posts: posts, isSuccessSubscribed: true));
+      emit(UserSubscribed());
+
+      emit(current.copyWith(posts: posts));
     } catch (e, st) {
-      emit(current.copyWith(isSuccessSubscribed: false));
+      emit(UserSubscribingFailure(error: e));
+      emit(current);
       talker.handle(e, st);
       errorHandler.handle(e);
     }
